pyproject.toml
[tool.poetry]
name = "delete_simplelogin_contacts"
version = "0.1.0"
description = "Delete a SimpleLogin API user's all contacts."
authors = ["solzard <29563481+solzard@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
requests = "^2.27.1"
tqdm = "^4.64.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
